resolving remaining requested changes
